Moving back InferenceRunnerConfig to BulkInferenceRunnerConfig
